refactor: alakítsa át az S-méter konverziót Hamlib STRENGTH értékekhez

+++ src/store/rig.ts
@@ -283,7 +283,8 @@ export const useRigStore = defineStore('rig', {
         // Get Signal Strength (S-meter)
         const strengthResponse = await rigctldService.getStrength();
         if (strengthResponse.success && strengthResponse.data) {
-          this.rigState.signalStrength = parseFloat(strengthResponse.data[0]) || 0;
+          // Store the raw Hamlib STRENGTH value (0-255)
+          this.rigState.signalStrength = parseInt(strengthResponse.data[0]) || 0;
         }
 
         this.lastUpdate = new Date();
